+++ Content.Goobstation.Server/IoC/ServerGoobContentIoC.cs
@@ -6,8 +6,10 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+using Content.Goobstation.Common.Antag;
 using Content.Goobstation.Common.JoinQueue;
 using Content.Goobstation.Common.MisandryBox;
+using Content.Goobstation.Server.Antag;
 using Content.Goobstation.Server.JoinQueue;
 using Content.Goobstation.Server.MisandryBox.JumpScare;
 using Content.Goobstation.Server.MisandryBox.Spider;
@@ -28,6 +30,7 @@ internal static class ServerGoobContentIoC
         instance.Register<IVoiceChatServerManager, VoiceChatServerManager>();
         instance.Register<IJoinQueueManager, JoinQueueManager>();
         instance.Register<IFullScreenImageJumpscare, ServerFullScreenImageJumpscare>();
+        instance.Register<ILastRolledAntagManager, LastRolledAntagManager>();
         instance.Register<ISpiderManager, ServerSpiderManager>();
     }
 }
