+++ Content.Server/Entry/EntryPoint.cs
@@ -131,7 +131,6 @@
 //
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-using Content.Server._Goobstation.Antag;
 using Content.Server._durkcode.ServerCurrency;
 using Content.Server.Acz;
 using Content.Server.Administration;
@@ -184,7 +183,6 @@ namespace Content.Server.Entry
         private IConnectionManager? _connectionManager;
 
         private ServerCurrencyManager? _currencyManager; // Goobstation
-        private LastRolledAntagManager? _lastAntagManager; // Goobstation
 
         /// <inheritdoc />
         public override void Init()
@@ -255,8 +253,6 @@ namespace Content.Server.Entry
                 IoCManager.Resolve<PlayerRateLimitManager>().Initialize();
                 _currencyManager = IoCManager.Resolve<ServerCurrencyManager>(); // Goobstation
                 _currencyManager.Initialize(); // Goobstation
-                _lastAntagManager = IoCManager.Resolve<LastRolledAntagManager>(); // Goobstation
-                _lastAntagManager.Initialize(); // Goobstation
             }
         }
