+++ Content.Server/IoC/ServerContentIoC.cs
@@ -163,7 +163,6 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 using Content.Server._durkcode.ServerCurrency;
-using Content.Server._Goobstation.Antag;
 using Content.Server._RMC14.LinkAccount; // RMC - Patreon
 using Content.Server.Administration;
 using Content.Server.Administration.Logs;
@@ -245,7 +244,6 @@ namespace Content.Server.IoC
             IoCManager.Register<MultiServerKickManager>();
             IoCManager.Register<CVarControlManager>();
             IoCManager.Register<ServerCurrencyManager>(); // Goobstation - Server Currency
-            IoCManager.Register<LastRolledAntagManager>(); // Goobstation - antag pity
             IoCManager.Register<LinkAccountManager>(); // RMC - Patreon
         }
     }
